Fix request proxy bug due to "Content-Length" header
This PR closes #97, an issue regarding specifying a custom proxy for
node-gcm requests.

Due to a bug with the `request` module (explained in
#97#issuecomment-162701641), the initial HTTP CONNECT request never
succeeds.

This commit works around the issue by:
1. Removing hard-coded declarations of Content-Type and Content-Length
headers (`request` will add them automatically if we supply JSON
payload)
2. Supply the request body in the `json` parameter (as an object - not
as a JSON-stringified string)
3. Remove the `parseAndRespond` function which is no longer needed, as
`request` returns a JSON object (it parses it for us)