Improved support for unicast messaging (sending to one device)
As per @alemonmk's suggestion in #187, when supplying a single
registration token inside an array, we should set the `to` property and
not the `registration_ids` property, so that the push notification will
be sent as a unicast message, as per the GCM documentation.

Added a check for a single-element array inside `sendNoRetry`.

The response from the GCM server is identical for unicast and multicast
messages.